pyproject.toml
[tool.poetry]
name = "Symfall"
version = "0.11.0"
description = "Simple messenger to improve self skills"
authors = ["swaroggot", "valeriiduz"]

[tool.poetry.dependencies]
python = "3.9.6"
packaging = "20.4"

psutil = "5.7.2"
sentry-sdk = "1.3.0"
psycopg2-binary = "2.8.6"

celery = "5.0.0"
redis = "3.5.3"

Django = "3.2.5"
drf-yasg = "1.20.0"
djangorestframework = "3.12.4"
django-extensions = "3.1.3"
django-health-check = "3.16.2"
django-cors-headers = "3.7.0"

django-storages = "1.11.1"
google-cloud-storage = "1.36.0"

Faker = "8.10.3"
Pillow = "8.2.0"

channels = "3.0.3"
channels-redis = "3.2.0"
django-filter = "2.4.0"

[tool.poetry.dev-dependencies]

black = {version = "21.7b0", allow-prereleases = true}
autoflake = "1.4"
isort = "5.6.2"
pylint = "2.9.5"
pylint-django = "2.4.4"

pytest-django = "4.4.0"
pytest-xdist = "2.3.0"
pytest-asyncio = "0.15.1"
pytest-cov = "2.12.1"
freezegun = "1.1.0"
coverage = "5.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


[tool.black]
line-length = 79
include = "\\.pyi?$"
